Fix capture of error when running gui in debug mode
When the gui is run in debug mode, the app is started twice. So, a port
is bound two times as well. However, the try except block did not catch
the error that the port was already in use (the first time the port was
bound), so the local webserver could not start. This was fixed by
capturing the generic Exception.