Move check import recipy before patched modules to log_init
The check whether recipy is imported before patched modules was moved
from multiple_insert to log_init. Log_init makes more sense, because we
can make sure a new run always exists when the check is done. It does,
however, require additional bookkeeping of which modules have patches.
This should be moved to a __init__ of the patch object. Also, the list
of patched modules is now stored in memory, this can be moved to the
database.