Store list of patched modules in database
The list is stored in a separate table called 'patches'.
The list of patches is reset every time recipy is imported.

Next step is to turn the added functionality into testable functions and
add tests.
Also, it might be an idea to store all information about a patch (i.e.,
add input functions and output functions) so, the user can inspect those
through the cli and gui.