Replace some cli search options with listsearch
Because it is not completely clear to me why there are three different
search options in the cli, and what exactly are the differences between
the search options, I just replaced two search options with the new
listsearch. I expect the different cli search options can (and should
be) merged.