Add new patch type that allows for multiple wrappers
PatchSimple allows for exactly one wrapper for input and output functions
respectively. For xarray output function save_mfdataset, the default wrapper
which logs argument 0 does not work, because it should be argument 1
that is logged. So, a second wrapper is required to be able to log this
function. Therefore a new type of patch was added.