Fix error with log_warnings coming from pandas Panel tests
Some of the tests using the pandas Panel object were failing. This is because they gave a
warning when used (as Panel is deprecated) and the log_warning function was failing. This is
because the definition of the function has changed between Python 2 and 3, and the previous
definition wasn't valid. This has now been replaced with a definition that works for both 2
and 3, therefore fixing the problem.