Need to create two joblib files for the mnist dataset for use in 2.7
So we now have checking for the python version and using the correct
sample data for each version.

However in doing this realised that format strings are not available
pre 3.6 so written them in .format notation.

Finally tensorflow is not available for python 2.7 on windows!
A bit random but means that to create data for travis I will need
to create a 2.7 environment on my macbook.